feat: monorepo phase 4b -- drift-session capability package (session/outcome/reward, ADR-100)
- Extract session.py, session_handover.py, session_queue_log.py, session_writer_lock.py,
  outcome_tracker.py, reward_chain.py and outcome_ledger/ into packages/drift-session
- Convert src/drift session and outcome files to compat re-export stubs
- Register drift-session as workspace member in pyproject.toml

(cherry picked from commit 307a55bbe233d391340c684f2dad9ce5cf496f42)
