+++ packages/drift-session/src/drift_session/__init__.py
@@ -1,0 +1,16 @@
+"""Session orchestration surface for Drift."""
+
+from drift_session.outcome_tracker import Outcome, OutcomeTracker, compute_fingerprint
+from drift_session.reward_chain import RewardScore, compute_reward
+from drift_session.session import DriftSession, OrchestrationMetrics, SessionManager
+
+__all__ = [
+    "DriftSession",
+    "OrchestrationMetrics",
+    "Outcome",
+    "OutcomeTracker",
+    "RewardScore",
+    "SessionManager",
+    "compute_fingerprint",
+    "compute_reward",
+]
