+++ packages/drift-session/src/drift_session/outcome_ledger/__init__.py
@@ -1,0 +1,28 @@
+"""Outcome-Feedback-Ledger (ADR-088 / K2 MVP)."""
+
+from __future__ import annotations
+
+from drift_session.outcome_ledger._models import (
+    LEDGER_SCHEMA_VERSION,
+    STALENESS_HISTORICAL_DAYS,
+    STALENESS_WARNING_DAYS,
+    AuthorType,
+    MergeTrajectory,
+    RecommendationOutcome,
+    TrajectoryDirection,
+)
+from drift_session.outcome_ledger.ledger_io import append_trajectory, load_trajectories
+from drift_session.outcome_ledger.reporter import render_markdown_report
+
+__all__ = [
+    "LEDGER_SCHEMA_VERSION",
+    "STALENESS_HISTORICAL_DAYS",
+    "STALENESS_WARNING_DAYS",
+    "AuthorType",
+    "MergeTrajectory",
+    "RecommendationOutcome",
+    "TrajectoryDirection",
+    "append_trajectory",
+    "load_trajectories",
+    "render_markdown_report",
+]
