+++ src/drift/outcome_ledger/__init__.py
@@ -1,28 +1,37 @@
-"""Outcome-Feedback-Ledger (ADR-088 / K2 MVP)."""
+"""Re-export stub -- drift_session.outcome_ledger (ADR-100 Phase 4b)."""
 
-from __future__ import annotations
+import importlib as _importlib
+import sys as _sys
 
-from drift.outcome_ledger._models import (
-    LEDGER_SCHEMA_VERSION,
-    STALENESS_HISTORICAL_DAYS,
-    STALENESS_WARNING_DAYS,
-    AuthorType,
-    MergeTrajectory,
-    RecommendationOutcome,
-    TrajectoryDirection,
+from drift_session.outcome_ledger import (  # noqa: F401
+    LEDGER_SCHEMA_VERSION as LEDGER_SCHEMA_VERSION,
 )
-from drift.outcome_ledger.ledger_io import append_trajectory, load_trajectories
-from drift.outcome_ledger.reporter import render_markdown_report
+from drift_session.outcome_ledger import (
+    STALENESS_HISTORICAL_DAYS as STALENESS_HISTORICAL_DAYS,
+)
+from drift_session.outcome_ledger import (
+    STALENESS_WARNING_DAYS as STALENESS_WARNING_DAYS,
+)
+from drift_session.outcome_ledger import (
+    AuthorType as AuthorType,
+)
+from drift_session.outcome_ledger import (
+    MergeTrajectory as MergeTrajectory,
+)
+from drift_session.outcome_ledger import (
+    RecommendationOutcome as RecommendationOutcome,
+)
+from drift_session.outcome_ledger import (
+    TrajectoryDirection as TrajectoryDirection,
+)
+from drift_session.outcome_ledger import (
+    append_trajectory as append_trajectory,
+)
+from drift_session.outcome_ledger import (
+    load_trajectories as load_trajectories,
+)
+from drift_session.outcome_ledger import (
+    render_markdown_report as render_markdown_report,
+)
 
-__all__ = [
-    "LEDGER_SCHEMA_VERSION",
-    "STALENESS_HISTORICAL_DAYS",
-    "STALENESS_WARNING_DAYS",
-    "AuthorType",
-    "MergeTrajectory",
-    "RecommendationOutcome",
-    "TrajectoryDirection",
-    "append_trajectory",
-    "load_trajectories",
-    "render_markdown_report",
-]
+_sys.modules[__name__] = _importlib.import_module("drift_session.outcome_ledger")
