+++ src/drift/outcome_ledger/_models.py
@@ -1,56 +1,28 @@
-"""Immutable Pydantic-Modelle fuer den Outcome-Feedback-Ledger (ADR-088)."""
+"""Re-export stub -- drift_session.outcome_ledger._models (ADR-100 Phase 4b)."""
 
-from __future__ import annotations
+import importlib as _importlib
+import sys as _sys
 
-from enum import StrEnum
+from drift_session.outcome_ledger._models import (
+    LEDGER_SCHEMA_VERSION as LEDGER_SCHEMA_VERSION,
+)
+from drift_session.outcome_ledger._models import (
+    STALENESS_HISTORICAL_DAYS as STALENESS_HISTORICAL_DAYS,
+)
+from drift_session.outcome_ledger._models import (
+    STALENESS_WARNING_DAYS as STALENESS_WARNING_DAYS,
+)
+from drift_session.outcome_ledger._models import (
+    AuthorType as AuthorType,
+)
+from drift_session.outcome_ledger._models import (
+    MergeTrajectory as MergeTrajectory,
+)
+from drift_session.outcome_ledger._models import (
+    RecommendationOutcome as RecommendationOutcome,
+)
+from drift_session.outcome_ledger._models import (  # noqa: F401
+    TrajectoryDirection as TrajectoryDirection,
+)
 
-from pydantic import BaseModel, ConfigDict, Field
-
-LEDGER_SCHEMA_VERSION: int = 1
-STALENESS_WARNING_DAYS: int = 90
-STALENESS_HISTORICAL_DAYS: int = 180
-
-
-class TrajectoryDirection(StrEnum):
-    IMPROVED = "improved"
-    REGRESSED = "regressed"
-    NEUTRAL = "neutral"
-    INDETERMINATE = "indeterminate"
-
-
-class AuthorType(StrEnum):
-    HUMAN = "human"
-    AI = "ai"
-    MIXED = "mixed"
-
-
-class RecommendationOutcome(BaseModel):
-    model_config = ConfigDict(frozen=True, extra="forbid")
-
-    recommendation_fingerprint: str = Field(min_length=1)
-    signal_type: str = Field(min_length=1)
-    task_kind: str | None = None
-    expected_delta: float | None = None
-    observed_delta: float
-    resolved: bool
-    correlation_confidence: float = Field(ge=0.0, le=1.0)
-    file_paths: tuple[str, ...] = Field(default_factory=tuple)
-
-
-class MergeTrajectory(BaseModel):
-    model_config = ConfigDict(frozen=True, extra="forbid")
-
-    schema_version: int = Field(default=LEDGER_SCHEMA_VERSION, ge=1)
-    merge_commit: str = Field(min_length=7)
-    parent_commit: str = Field(min_length=7)
-    timestamp: str
-    author_type: AuthorType
-    ai_attribution_confidence: float = Field(ge=0.0, le=1.0)
-    pre_score: float
-    post_score: float
-    delta: float
-    direction: TrajectoryDirection
-    per_signal_delta: dict[str, float] = Field(default_factory=dict)
-    recommendation_outcomes: tuple[RecommendationOutcome, ...] = Field(default_factory=tuple)
-    staleness_days: int = Field(default=0, ge=0)
-    notes: tuple[str, ...] = Field(default_factory=tuple)
+_sys.modules[__name__] = _importlib.import_module("drift_session.outcome_ledger._models")
